% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_lookup.R
\name{location_lookup}
\alias{location_lookup}
\title{Function for doing lookup of place-names and return closest matching bus-stops}
\usage{
location_lookup(location, api_key, max_locations = 5L)
}
\arguments{
\item{location}{String to search for.}

\item{api_key}{Key for the api.}

\item{max_locations}{Maximum number of matches to return, default = 5, 
accepted range 1-50.}
}
\value{
data.frame with stop-Id, extId, name, longitude and latitude (WGS84),
 weight (amount of trafic) and products (transportations available)
}
\description{
Function for doing lookup of place-names and return closest matching bus-stops
}
\references{
https://www.trafiklab.se/api/resrobot-reseplanerare/platsuppslag
}
