% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearby_stops_app.R
\name{nearby_stops_app}
\alias{nearby_stops_app}
\title{Function for running shiny app based on nearby stops API}
\usage{
nearby_stops_app()
}
\value{
nothing...
}
\description{
Function for running shiny app based on nearby stops API
}
